+++ plugins/domains/dailymotion.com/dailymotion.com.js
@@ -13,20 +13,12 @@ module.exports = {
         "oembed-video-responsive"
     ],
 
-    getLink: function (url, cb) { 
-
-        if (url.match(/^https?:\/\/www\.dailymotion\.com\/swf\/video\//i)) {
-            cb ({
-                redirect: url.replace(/\/swf/, '')
-            })
-            
-        } else {
-            cb(null, {
-                href: "http://static1.dmcdn.net/images/apple-touch-icon.png.vcbf86c6fe83fbbe11",
-                type: CONFIG.T.image_icon,
-                rel: CONFIG.R.icon
-            });
-        }
+    getLink: function (url, oembed, cb) {
+        cb(null, {
+            href: "http://static1.dmcdn.net/images/apple-touch-icon.png.vcbf86c6fe83fbbe11",
+            type: CONFIG.T.image_icon,
+            rel: CONFIG.R.icon
+        });
     },
 
     tests: [