+++ plugins/domains/facebook.com/facebook.post.js
@@ -26,10 +26,11 @@ module.exports = {
     getLink: function(facebook_post, options) {
         return {
             type: CONFIG.T.text_html,
-            rel: [CONFIG.R.app, CONFIG.R.inline, CONFIG.R.ssl],
+            rel: [CONFIG.R.app, CONFIG.R.ssl, CONFIG.R.html5],
             template_context: {
                 title: facebook_post.title,
                 url: facebook_post.url,
+                type: 'fb-post',
                 width: options.maxWidth || DEFAULT_WIDTH
             },
             width: options.maxWidth || DEFAULT_WIDTH