+++ plugins/links/sailthru.js
@@ -17,7 +17,10 @@ module.exports = {
         return [].concat(
             utils.getImageLink('sailthru.thumb', meta) || [],
             utils.getImageLink('sailthru.image.full', meta) || [],
-            utils.getImageLink('sailthru.image.thumb', meta) || []
+            utils.getImageLink('sailthru.image.thumb', meta) || [],
+            utils.getImageLink('sailthru.playimage', meta) || [],
+            utils.getImageLink('sailthru.author-avatar', meta) || [],
+            utils.getImageLink('sailthru.brandthumbimage', meta) || []
         );
     }
 };